Renamed commands package to command

Took 3 minutes
