+++ src/main/kotlin/microservice/dungeon/game/aggregates/command/domain/CommandObject.kt
@@ -1,4 +1,4 @@
-package microservice.dungeon.game.aggregates.commands.domain
+package microservice.dungeon.game.aggregates.command.domain
 
 import java.util.*
 import javax.persistence.Embeddable
