+++ src/main/kotlin/microservice/dungeon/game/aggregates/command/dtos/CommandDTO.kt
@@ -1,0 +1,17 @@
+package microservice.dungeon.game.aggregates.command.dtos
+
+import microservice.dungeon.game.aggregates.command.domain.CommandObject
+import microservice.dungeon.game.aggregates.command.domain.CommandType
+import java.util.*
+
+class CommandDTO(
+    val gameId: UUID,
+
+    val playerId: UUID,
+
+    val robotId: UUID,
+
+    val commandType: CommandType,
+
+    val commandObject: CommandObject
+)
