+++ src/main/kotlin/microservice/dungeon/game/aggregates/command/repositories/RoundCommandsRepository.kt
@@ -1,6 +1,6 @@
-package microservice.dungeon.game.aggregates.commands.repositories
+package microservice.dungeon.game.aggregates.command.repositories
 
-import microservice.dungeon.game.aggregates.commands.domain.RoundCommands
+import microservice.dungeon.game.aggregates.command.domain.RoundCommands
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
